% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapl.R
\name{lapl_context}
\alias{lapl_context}
\title{Return context: full source of this client}
\usage{
lapl_context()
}
\value{
character: source code of this file
}
\description{
Return context: full source of this client
}
