% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapl.R
\name{lapl_hot_titles}
\alias{lapl_hot_titles}
\title{Fetch hot/popular titles across all categories}
\usage{
lapl_hot_titles()
}
\value{
tibble: feed_id, title, link, author, genre, pub_date, cover_url
}
\description{
Fetch hot/popular titles across all categories
}
