% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapl.R
\name{lapl_list}
\alias{lapl_list}
\title{List available LAPL book feeds}
\usage{
lapl_list(category = NULL)
}
\arguments{
\item{category}{Optional filter: "new_titles" or "hot_titles"}
}
\value{
tibble: id, name, url, category
}
\description{
List available LAPL book feeds
}
