% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapl.R
\name{lapl_new_titles}
\alias{lapl_new_titles}
\title{Fetch new titles across all genres}
\usage{
lapl_new_titles()
}
\value{
tibble: feed_id, title, link, author, genre, pub_date, cover_url
}
\description{
Fetch new titles across all genres
}
