% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapl.R
\name{lapl_search}
\alias{lapl_search}
\title{Search LAPL titles by keyword}
\usage{
lapl_search(query, feeds = NULL)
}
\arguments{
\item{query}{Search string (case-insensitive, matches title or author)}

\item{feeds}{Which feeds to search (default: all). Vector of feed ids.}
}
\value{
tibble: feed_id, title, link, author, genre, pub_date, cover_url
}
\description{
Search LAPL titles by keyword
}
