% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapl.R
\name{lapl_titles}
\alias{lapl_titles}
\title{Fetch titles from an LAPL book feed}
\usage{
lapl_titles(feed_id)
}
\arguments{
\item{feed_id}{Feed id from lapl_list() (e.g. "new_scifi", "hot_fiction")}
}
\value{
tibble: title, link, author, genre, pub_date, cover_url
}
\description{
Fetch titles from an LAPL book feed
}
